Creates a generic Measurement as a super class for Length

+++ src/problem_3/model/VolumeUnit.java
@@ -9,7 +9,7 @@ public enum VolumeUnit {
     this.standardValue = standardValue;
   }
 
-  public double normalize(double volume) {
+  public double normalize(int volume) {
     return this.standardValue * volume;
   }
 }
